% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_par_run.R
\name{car_par_run}
\alias{car_par_run}
\title{Parallelize many simulations}
\usage{
car_par_run(rho, dims, prob_blue, trials, replicates, cores = 1)
}
\arguments{
\item{rho}{Either a proportion from 0 to 1 of the grid to be occupied by
cars, or an integer value specifying the number of cars on the grid.}

\item{dims}{A numeric vector of length 2 specifying the number of
(rows, columns) of the grid.}

\item{prob_blue}{Probability that a given car is blue.}

\item{trials}{The number of moves that the cars will make, blue cars move on
odd time periods, red cars move on even time periods.}

\item{replicates}{The number of different simulations.}

\item{cores}{The number of cores this process will be run on.}
}
\value{
A list of "carsimr_list" objects.
}
\description{
This function allows users to simulate car movements on many different grids
at the same time. The user can specify a number of cores greater than 1 if
they would like to parallelize this process.
}
