% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{carsimr}
\alias{carsimr}
\title{Construct carsimr objects}
\usage{
carsimr(x)
}
\arguments{
\item{x}{A matrix to be turned into an object of type carsimr.}
}
\value{
An object of type carsimr, essentially a matrix with a generic for
plot.
}
\description{
This function constructs carsimr objects.
}
