% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{carsimr_list}
\alias{carsimr_list}
\title{Construct carsimr_list objects}
\usage{
carsimr_list(x)
}
\arguments{
\item{x}{A list to be turned into an object of type carsimr_list.}
}
\value{
An object of type carsimr_list, essentially a list of matrices with
a generic for plot.
}
\description{
This function constructs carsimr_list objects.
}
