% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{cpp_convert_carsimr}
\alias{cpp_convert_carsimr}
\title{Convert objects returned from Rcpp to usable R classes}
\usage{
cpp_convert_carsimr(x)
}
\arguments{
\item{x}{A matrix returned from initialize_grid_cpp.}
}
\value{
A carsimr object.
}
\description{
This function would construct carsimr objects, after they have been returned
from initialize_grid_cpp.
}
