% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_grid.R
\name{initialize_grid}
\alias{initialize_grid}
\title{Initializing the grid}
\usage{
initialize_grid(rho, dims, prob_blue)
}
\arguments{
\item{rho}{Either a proportion from 0 to 1 of the grid to be occupied by
cars, or an integer value specifying the number of cars on the grid.}

\item{dims}{A numeric vector of length 2 specifying the number of
(rows, columns) of the grid.}

\item{prob_blue}{Probability that a given car is blue.}
}
\value{
An object of class carsimr, essentially a matrix with a generic for
plot.
}
\description{
This function initializes a matrix, representing a grid of cars.
}
\examples{
# Proportion rho
initialize_grid(rho = 0.5, dims = c(3, 5), prob_blue = 0.5)

# Integer rho
initialize_grid(rho = 8, dims = c(3, 5), prob_blue = 0.5)

}
