% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_cars.R
\name{move_cars}
\alias{move_cars}
\title{Moving cars}
\usage{
move_cars(t_grid, trials)
}
\arguments{
\item{t_grid}{An object of class carsimr.}

\item{trials}{The number of moves that the cars will make, blue cars move on
odd time periods, red cars move on even time periods.}
}
\value{
An object of class carsimr_list, which contains the initial_grid
along with along with each of the "trial" movements.
}
\description{
This function moves the cars in the grid created by the initialize_grid
function.
}
