% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.carsimr}
\alias{plot.carsimr}
\title{Plot a carsimr object}
\usage{
\method{plot}{carsimr}(x, y, ...)
}
\arguments{
\item{x}{An object of class carsimr.}

\item{y}{Required to create plot generic, not used.}

\item{...}{Additional parameters to be passed into the plot function.}
}
\value{
Plots the state of the grid.
}
\description{
This function allows users to plot an object of class carsimr, where the
blue cars are represented by blue tiles on a matrix, and the red cars are
represented by red tiles on a matrix.
}
