% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.carsimr_list}
\alias{plot.carsimr_list}
\title{Sequential plots of carsimr objects}
\usage{
\method{plot}{carsimr_list}(x, y, pause = 1, ...)
}
\arguments{
\item{x}{An object of class carsimr_list.}

\item{y}{Required to create plot generic, not used.}

\item{pause}{Length of time (in seconds) in between plotting objects of
carsimr.}

\item{...}{Additional parameters to be passed into plot.}
}
\value{
A series of plots representing the movement of cars along the grid
for iterations of time.
}
\description{
This function allows users to plot an object of class carsimr_list, which
successively plots carsimr objects
}
