% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_blocks.R
\name{track_blocks}
\alias{track_blocks}
\title{Track blocked cars}
\usage{
track_blocks(carsimr_list)
}
\arguments{
\item{carsimr_list}{An object of type carsimr_list}
}
\value{
A list of length 2, where the first element is a vector of block
proportions for BLUE cars on blue moves, and the second element is a vector
of block proportions for RED cars on red moves
}
\description{
This function finds the proportion of blocked cars at each time iteration
for a carsimr_list objects.
}
